feat(SingleSessionManager): store streams of an identifier in an array to prevent unexpected ending
For every identifier, streams are now stored in arrays; every time a session overrides, its stream
will be pushed to the top of the array, and popped after the stream is ended. The original one (the
one that is overrided) will stop receiving messages in this period of time, instead of being
force-ended.